adding fieldRouter
